feat: adjust all model input and output so that model can predict successfully

+++ models/weather.go
@@ -4,16 +4,16 @@ import "encoding/json"
 
 type Weather struct {
 	Date string  `json:"date"` // Date in YYYY-MM-DD format
-	Tavg float64 `json:"tavg"` // Average temperature in °C
-	TMin float64 `json:"tmin"` // Minimum temperature in °C
-	TMax float64 `json:"tmax"` // Maximum temperature in °C
-	Prcp float64 `json:"prcp"` // Precipitation in mm
-	Snow float64 `json:"snow"` // Snowfall in mm
-	Wdir float64 `json:"wdir"` // Wind direction in degrees
-	Wspd float64 `json:"wspd"` // Wind speed in km/h
-	Wpgt float64 `json:"wpgt"` // Peak wind gust in km/h
-	Pres float64 `json:"pres"` // Air pressure in hPa
-	Tsun float64 `json:"tsun"` // Sunshine duration in minutes
+	Tavg float32 `json:"tavg"` // Average temperature in °C
+	TMin float32 `json:"tmin"` // Minimum temperature in °C
+	TMax float32 `json:"tmax"` // Maximum temperature in °C
+	Prcp float32 `json:"prcp"` // Precipitation in mm
+	Snow float32 `json:"snow"` // Snowfall in mm
+	Wdir float32 `json:"wdir"` // Wind direction in degrees
+	Wspd float32 `json:"wspd"` // Wind speed in km/h
+	Wpgt float32 `json:"wpgt"` // Peak wind gust in km/h
+	Pres float32 `json:"pres"` // Air pressure in hPa
+	Tsun float32 `json:"tsun"` // Sunshine duration in minutes
 }
 
 type WeatherResponse struct {
